refactor: Move fetching kakao user data codes to utils.py
